+++ src/main/typescript/ruleEngine/ruleElements.ts
@@ -1,18 +1,19 @@
+import type { ValidateFunction } from "ajv";
+import type { JTDErrorObject } from "ajv/dist/jtd";
 import { getGame } from "../foundryHelpers.js";
 import type WvItem from "../item/wvItem.js";
-import type {
-  RuleElementLike,
-  KnownRuleElementSource,
-  UnknownRuleElementSource,
-  RuleElementTarget
-} from "./ruleElement.js";
-import FlatModifier from "./ruleElements/flatModifier.js";
-import RuleElementMessage from "./ruleElementMessage.js";
+import AdditionalPropMessage from "./messages/additionalPropMessage.js";
 import MissingPropMessage from "./messages/missingPropMessage.js";
 import WrongTypeMessage from "./messages/wrongTypeMessage.js";
-import validate from "../validators/ruleElementSource.js";
-import type { ErrorObject } from "ajv";
-import { isValidTarget } from "./ruleElement.js";
+import type {
+  KnownRuleElementSource,
+  RuleElementLike,
+  UnknownRuleElementSource
+} from "./ruleElement.js";
+import RuleElementMessage from "./ruleElementMessage.js";
+import FlatModifier from "./ruleElements/flatModifier.js";
+import type RuleElementSource from "./ruleElementSource.js";
+import { createValidator } from "./ruleElementSource.js";
 
 /** RuleElement identifier strings */
 export const RULE_ELEMENT_IDS = {
@@ -33,6 +34,10 @@ export type MappedRuleElementId = keyof typeof RULE_ELEMENTS;
  * A factory class for RuleElements.
  */
 export default class RuleElements {
+  /** A cached validator function for rule element sources. */
+  protected static isValidRuleElementSource: ValidateFunction<RuleElementSource> | null =
+    null;
+
   /**
    * Create a new RuleElementSource, suitable for when the user just added a
    * new effect and has not filled out the data yet.
@@ -63,64 +68,68 @@ export default class RuleElements {
   ): RuleElementLike {
     const messages: RuleElementMessage[] = [];
 
+    const validate = RuleElements.validate;
+
     // Check the passed JSON source against the schema. When it is invalid, only
     // return a RuleElementLike.
     if (!validate(source)) {
-      validate.errors?.forEach((error) =>
-        messages.push(this.translateError(error))
-      );
+      for (const error of validate.errors as JTDErrorObject[]) {
+        messages.push(this.translateError(error));
+      }
 
       return { item, messages, source };
     }
 
-    // Check if the target of the RuleElementSource is valid. If not, return a
-    // RuleElementLike.
-    let target: RuleElementTarget;
-    if (!isValidTarget(source.target)) {
-      messages.push(
-        new RuleElementMessage(
-          "wv.ruleEngine.errors.semantic.unknownTarget",
-          "error"
-        )
-      );
-
-      return { item, messages, source };
-    } else {
-      target = source.target;
-    }
-
-    // Check if the type of RuleElement is known. If not, return a
-    // RuleElementLike.
-    let type: MappedRuleElementId;
-    if (!isMappedRuleElementType(source.type)) {
-      messages.push(
-        new RuleElementMessage(
-          "wv.ruleEngine.errors.semantic.unknownRuleElement",
-          "error"
-        )
-      );
-
-      return { item, messages, source };
-    } else {
-      type = source.type;
-    }
-
+    const target = source.target;
+    const type = source.type;
     return new RULE_ELEMENTS[type]({ ...source, target, type }, item, messages);
   }
 
-  /** Translate an AJV ErrorObject to a RuleElementMessage. */
-  protected static translateError(error: ErrorObject): RuleElementMessage {
+  /** Translate an AJV JTDErrorObject to a RuleElementMessage. */
+  protected static translateError(error: JTDErrorObject): RuleElementMessage {
     switch (error.keyword) {
-      case "required":
-        return new MissingPropMessage(
-          error.instancePath,
-          error.params.missingProperty
-        );
+      case "properties":
+        if ("additionalProperty" in error.params) {
+          return new AdditionalPropMessage(
+            error.instancePath,
+            error.params.additionalProperty
+          );
+        } else if ("missingProperty" in error.params) {
+          return new MissingPropMessage(
+            error.instancePath,
+            error.params.missingProperty
+          );
+        }
+        break;
+
       case "type":
         return new WrongTypeMessage(error.instancePath, error.params.type);
-      default:
-        return new RuleElementMessage("wv.ruleEngine.errors.semantic.unknown");
+
+      case "enum":
+        switch (error.schemaPath) {
+          case "/properties/target/enum":
+            return new RuleElementMessage(
+              "wv.ruleEngine.errors.semantic.unknownTarget",
+              "error"
+            );
+          case "/properties/type/enum":
+            return new RuleElementMessage(
+              "wv.ruleEngine.errors.semantic.unknownRuleElement",
+              "error"
+            );
+        }
     }
+
+    console.dir(error);
+    return new RuleElementMessage("wv.ruleEngine.errors.semantic.unknown");
+  }
+
+  /** Get a possibly cached validation function. */
+  protected static get validate(): ValidateFunction<RuleElementSource> {
+    if (RuleElements.isValidRuleElementSource)
+      return RuleElements.isValidRuleElementSource;
+
+    return (RuleElements.isValidRuleElementSource = createValidator());
   }
 }